+++ src/Components/Bucket/BucketCard.tsx
@@ -5,16 +5,17 @@ import {BucketInfo} from "reduct-js";
 
 // @ts-ignore
 import prettierBytes from "prettier-bytes";
-import {DeleteOutlined} from "@ant-design/icons";
+import {DeleteOutlined, SettingOutlined} from "@ant-design/icons";
 
 interface Props {
     bucket: BucketInfo;
     index: number;
     onRemove: (name: string) => void;
+    onSettings: (name: string) => void;
 }
 
 export default function BucketCard(props: Readonly<Props>) {
-    const [confirm, setConfirm] = useState(false);
+    const [visible, setVisible] = useState(false);
     const {bucket, index} = props;
 
     const n = (big: BigInt) => {
@@ -27,9 +28,18 @@ export default function BucketCard(props: Readonly<Props>) {
             {largest: 1});
     };
 
+    const onOk = () => {
+        props.onRemove(bucket.name);
+        setVisible(false);
+    };
+
     return (<Card key={index} id={bucket.name} title={bucket.name} style={{margin: "0.5em"}}
                   actions={[
-                      <DeleteOutlined title="Remove" onClick={() => setConfirm(true)}/>
+                      <SettingOutlined title="Settings" onClick={() => {
+                          props.onSettings(bucket.name);
+                      }}/>,
+                      <DeleteOutlined title="Remove" onClick={() => setVisible(true)}/>,
+
                   ]}>
         <Row gutter={16}>
             <Col span={8}>
@@ -42,7 +52,8 @@ export default function BucketCard(props: Readonly<Props>) {
                 <Statistic title="History" value={bucket.entryCount > 0n ? getHistory() : "---"}/>
             </Col>
         </Row>
-        <Modal visible={confirm} onOk={()=> props.onRemove(bucket.name)} onCancel={() => setConfirm(false)} closable={false} okText="remove" okType="danger">Remove <b>{bucket.name}</b>?</Modal>
+        <Modal visible={visible} onOk={onOk} onCancel={() => setVisible(false)} closable={false} okText="Remove"
+               okType="danger">Remove <b>{bucket.name}</b>?</Modal>
     </Card>);
 
 }