+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 **Features**:
 
 * Use Reduct's colours for UI, [PR-7](https://github.com/reduct-storage/web-console/pull/7)
-
+* Implement changing bucket settings, [PR-9](https://github.com/reduct-storage/web-console/pull/9)
 **Other**:
 
 * Update reduct-js to 0.4.0, [PR-6](https://github.com/reduct-storage/web-console/pull/6)